Finish calculator main logic
After a lot of rework on the code, the main logic of the calculator is finished, only the button "." that doesn't work yet. A lot of the code was changed and functions like firstNumFn and secondNumFn were erased to make the code more simpler and effective, as well some new variables were created to help on the operations, there are some comments explaining the new changes and some new variables.